refactor: move endpoint queries to the queries folder
Part of #1169
